Refactor datasource list component to a seperate file
