+++ app/client/src/components/editorComponents/form/fields/EmbeddedDatasourcePathField.tsx
@@ -38,7 +38,7 @@ import { DATA_SOURCES_EDITOR_ID_URL } from "constants/routes";
 import Icon, { IconSize } from "components/ads/Icon";
 import Text, { TextType } from "components/ads/Text";
 import history from "utils/history";
-import { getDatasourceInfo } from "pages/Editor/APIEditor/Form";
+import { getDatasourceInfo } from "pages/Editor/APIEditor/DatasourceList";
 import * as FontFamilies from "constants/Fonts";
 
 type ReduxStateProps = {
